refactor: 챌린지 미리보기 카드 공통 프레젠테이션 뷰 추출
Create/Edit 미리보기 카드의 동일한 표시(chrome)를 ChallengePreviewCardView
로 추출하고, 두 카드는 폼 값 → 원시값 변환만 담당하는 얇은 어댑터로 남긴다.
출력/동작 완전 동일(중복 JSX ~130줄 제거).

의도적 범위 제한(로드맵 8, 리스크 높음):
Create/Edit 폼은 값 모델이 근본적으로 다르다 —
- Create: periodType/participationType/goalType enum + period/periodNumber
- Edit: isGroup/isFixedGoal boolean + isStarted 게이트 기반 필드 잠금
이 상태에서 섹션을 mode prop 으로 통합하려면 두 Zod 스키마부터 병합하고
edit 전용 잠금 정책을 분기로 이관해야 해, 잠금 정책 회귀 위험이 크고
mode 분기 mega-component 가 오히려 복잡해진다(공통 SectionCard 는 이미
공유 중). 따라서 값 모델과 무관한 "표시"만 안전하게 공통화했다.
